+++ src/platform/windows/gamepad.rs
@@ -57,6 +57,10 @@ impl Gilrs {
         }
     }
 
+    pub fn with_mappings(_sdl_mapping: &str) -> Self {
+        Self::new()
+    }
+
     pub fn next_event(&mut self) -> Option<(usize, Event)> {
         self.rx.try_recv().ok()
     }